compiler, runtime: replace hashmap code with Go 1.7 hashmap
This change removes the gccgo-specific hashmap code and replaces it with
the hashmap code from the Go 1.7 runtime.  The Go 1.7 hashmap code is
more efficient, does a better job on details like when to update a key,
and provides some support against denial-of-service attacks.

The compiler is changed to call the new hashmap functions instead of the
old ones.

The compiler now tracks which types are reflexive and which require
updating when used as a map key, and records the information in map type
descriptors.

Map_index_expression is simplified.  The special case for a map index on
the right hand side of a tuple expression has been unnecessary for some
time, and is removed.  The support for specially marking a map index as
an lvalue is removed, in favor of lowering an assignment to a map index
into a function call.  The long-obsolete support for a map index of a
pointer to a map is removed.

The __go_new_map_big function (known to the compiler as
Runtime::MAKEMAPBIG) is no longer needed, as the new runtime.makemap
function takes an int64 hint argument.

The old map descriptor type and supporting expression is removed.

The compiler was still supporting the long-obsolete syntax `m[k] = 0,
false` to delete a value from a map.  That is now removed, requiring a
change to one of the gccgo-specific tests.

The builtin len function applied to a map or channel p is now compiled
as `p == nil ? 0 : *(*int)(p)`.  The __go_chan_len function (known to
the compiler as Runtime::CHAN_LEN) is removed.

Support for a shared zero value for maps to large value types is
introduced, along the lines of the gc compiler.  The zero value is
handled as a common variable.

The hash function is changed to take a seed argument, changing the
runtime hash functions and the compiler-generated hash functions.
Unlike the gc compiler, both the hash and equal functions continue to
take the type length.

Types that can not be compared now store nil for the hash and equal
functions, rather than pointing to functions that throw.  Interface hash
and comparison functions now check explicitly for nil.  This matches the
gc compiler and permits a simple implementation for ismapkey.

The compiler is changed to permit marking struct and array types as
incomparable, meaning that they have no hash or equal function.  We use
this for thunk types, removing the existing special code to avoid
generating hash/equal functions for them.

The C runtime code adds memclr, memequal, and memmove functions.

The hashmap code uses go:linkname comments to make the functions
visible, as otherwise the compiler would discard them.

The hashmap code comments out the unused reference to the address of the
first parameter in the race code, as otherwise the compiler thinks that
the parameter escapes and copies it onto the heap.  This is probably not
needed when we enable escape analysis.

Several runtime map tests that ere previously skipped for gccgo are now
run.

The Go runtime picks up type kind information and stubs.  The type kind
information causes the generated runtime header file to define some
constants, including `empty`, and the C code is adjusted accordingly.

A Go-callable version of runtime.throw, that takes a Go string, is
added to be called from the hashmap code.

Change-Id: If8baa002241195f059be58067ad9a4b6f64ef669
Reviewed-on: https://go-review.googlesource.com/29447
Reviewed-by: Than McIntosh <thanm@google.com>
Reviewed-by: Ian Lance Taylor <iant@golang.org>