runtime: copy internal locking code from Go 1.7 runtime
Remove the old locking code written in C.

Add a shell script mkrsysinfo.sh to generate the runtime_sysinfo.go
file, so that we can get Go copies of the system time structures and
other types.

Tweak the compiler so that when compiling the runtime package the
address operator does not cause local variables to escape.  When the gc
compiler compiles the runtime, an escaping local variable is treated as
an error.  We should implement that, instead of this change, when escape
analysis is turned on.

Tweak the compiler so that the generated C header does not include names
that start with an underscore followed by a non-upper-case letter,
except for the special cases of _defer and _panic.  Otherwise we
translate C types to Go in runtime_sysinfo.go and then generate those Go
types back as C types in runtime.inc, which is useless and painful for
the C code.

Change entersyscall and friends to take a dummy argument, as the gc
versions do, to simplify calls from the shared code.

Change-Id: I983603e4d2915bbef19054636a2d2129c14c4e85
Reviewed-on: https://go-review.googlesource.com/30079
Reviewed-by: Than McIntosh <thanm@google.com>
Reviewed-by: Ian Lance Taylor <iant@golang.org>