libgo: strip most C macros from runtime.inc
The Go runtime package is picking up C macros from runtime_sysinfo.go
and then re-exporting them to runtime.inc.  This can cause name
conflicts.  Change the Makefile so that we only put the macros we need
into runtime.inc.  These are the constants that are actually defined by
Go code, not runtime_sysinfo.go.  There are only a few, so we can
pattern match.

This is an additional hack on runtime.inc.  The long term goal is to
convert the runtime package to Go and eliminate runtime.inc entirely, so
a few hacks seem acceptable.

Fixes GCC PR 77809.

Change-Id: If469118ddc984e08b652e37110a0353f65b3c2b5
Reviewed-on: https://go-review.googlesource.com/30167
Reviewed-by: Than McIntosh <thanm@google.com>
Reviewed-by: Ian Lance Taylor <iant@golang.org>