syscall, internal/syscall/unix: Fix getrandom, clone on sparc64
Since sparc is a valid architecture, the name of
getrandom_linux_sparc.go means that it will be ignored on sparc64,
even though it's whitelisted with a +build line.

On SPARC, clone has a unique return value convention which requires
some inline assembly to convert it to the normal convention.

Change-Id: I8c972bd003fec030447dc31363ddba74609f8816
Reviewed-on: https://go-review.googlesource.com/30873
Reviewed-by: Ian Lance Taylor <iant@golang.org>