runtime: copy mstats code from Go 1.7 runtime
This replaces mem.go and the C runtime_ReadMemStats function with the Go
1.7 mstats.go.

The GCStats code is commented out for now.  The corresponding gccgo code
is in runtime/mgc0.c.

The variables memstats and worldsema are shared between the Go code and
the C code, but are not exported.  To make this work, add temporary
accessor functions acquireWorldsema, releaseWorldsema, getMstats (the
latter known as mstats in the C code).

Check the preemptoff field of m when allocating and when considering
whether to start a GC.  This works with the new stopTheWorld and
startTheWorld functions in Go, which are essentially the Go 1.7
versions.

Change the compiler to stack allocate closures when compiling the
runtime package.  Within the runtime packages closures do not escape.
This is similar to what the gc compiler does, except that the gc
compiler, when compiling the runtime package, gives an error if escape
analysis shows that a closure does escape.  I added this here because
the Go version of ReadMemStats calls systemstack with a closure, and
having that allocate memory was causing some tests that measure memory
allocations to fail.

Change-Id: I711181717b719905950f3f002b061006e397563f
Reviewed-on: https://go-review.googlesource.com/30972
Reviewed-by: Ian Lance Taylor <iant@golang.org>