runtime: rewrite interface code into Go
I started to copy the Go 1.7 interface code, but the gc and gccgo
representations of interfaces are too different.  So instead I rewrote
the gccgo interface code from C to Go.  The code is largely the same as
it was, but the names are more like those used in the gc runtime.

I also copied over the string comparison functions, and tweaked the
compiler to use eqstring when comparing strings for equality.

Change-Id: I6cc5e16535f9a3a881080573294c0abc0ed9ca11
Reviewed-on: https://go-review.googlesource.com/31591
Reviewed-by: Than McIntosh <thanm@google.com>
Reviewed-by: Ian Lance Taylor <iant@golang.org>