compiler, runtime: copy slice code from Go 1.7 runtime
Change the compiler handle append as the gc compiler does: call a
function to grow the slice, but otherwise assign the new elements
directly to the final slice.

For the current gccgo memory allocator the slice code has to call
runtime_newarray, not mallocgc directly, so that the allocator sets the
TypeInfo_Array bit in the type pointer.

Rename the static function cnew to runtime_docnew, so that the stack
trace ignores it when ignoring runtime functions.  This was needed to
fix the runtime/pprof tests on 386.

Change-Id: I288d3177f8c191db7ce0d0448a20113f3306e4a0
Reviewed-on: https://go-review.googlesource.com/32218
Reviewed-by: Than McIntosh <thanm@google.com>