libgo: only rebuild package if dependent .gox has changed
Use mvifdiff and stamp files to track whether a .gox file has changed.
When package A depends on package B, and we rebuild package B, only
rebuild package A if package B's .gox file changes.  This is safe
because when package A imports package B it only reads package B's .gox
file.  This means that changes that do not affect export information
will not cause dependent packages to be recompiled.

Change-Id: If084363115f68293a400401ea18949c44d34118e
Reviewed-on: https://go-review.googlesource.com/32476
Reviewed-by: Than McIntosh <thanm@google.com>