runtime: copy signal code from Go 1.7 runtime
Add a little shell script to auto-generate runtime.sigtable from the
known signal names.

Force the main package to always import the runtime package.  Otherwise
some runtime package global variables may never be initialized.

Set the syscallsp and syscallpc fields of g when entering a syscall, so
that the runtime package knows when a g is executing a syscall.

Fix runtime.funcPC to avoid dead store elimination of the interface
value when the function is inlined.

Change-Id: Id3a932d344391bb6d9634760372565bd4e26057e
Reviewed-on: https://go-review.googlesource.com/33025
Reviewed-by: Than McIntosh <thanm@google.com>