compiler: rework static initializer code
Rename is_immutable to is_static_initializer to try to capture what it
really means.  Be more precise about when an address expression, or a
binary expression, can be a static initializer.  Don't check whether a
type has pointers when deciding whether an initializer must be
read-write, just check whether it is being used to initialize a global
variable.  To make that work set the Translate_context function to NULL
for a global variable with a static initializer.

The effect of this is to let more global variables be initialized
directly, rather than being initialized in the generated init function.

Change-Id: Ib0fa38a73968ce705f3b9e3101d5fadbfd93309d
Reviewed-on: https://go-review.googlesource.com/32917
Reviewed-by: Than McIntosh <thanm@google.com>