runtime: don't crash if signal handler info argument is nil
Apparently on Solaris 10 a SA_SIGINFO signal handler can be invoked with
a nil info argument.  I would not have believed it but I've now seen it
happen, and the sigaction man page actually says "If the second argument
is not equal to NULL, it points to a siginfo_t structure...."  So, if
that happens, don't crash.

Also fix another case where we want to make sure that &T{} does not
allocate.

Change-Id: I6479c89653aea5d6c1e0ee94a5cb96373f6395ad
Reviewed-on: https://go-review.googlesource.com/33150
Reviewed-by: Than McIntosh <thanm@google.com>