runtime, reflect: rewrite Go to FFI type conversion in Go
As we move toward the Go 1.7 garbage collector, it's essential that all
allocation of values that can contain Go pointers be done using the
correct type descriptor.  That is simplest if we do all such allocation
in Go code.  This rewrites the code that converts from a Go type to a
libffi CIF into Go.

Change-Id: I9d658121391a597c4664e76879d11969c07632c0
Reviewed-on: https://go-review.googlesource.com/33353
Reviewed-by: Than McIntosh <thanm@google.com>