runtime: rewrite panic/defer code from C to Go
The actual stack unwind code is still in C, but the rest of the code,
notably all the memory allocation, is now in Go.  The names are changed
to the names used in the Go 1.7 runtime, but the code is necessarily
somewhat different.

The __go_makefunc_can_recover function is dropped, as the uses of it
were removed in https://golang.org/cl/198770044.

Change-Id: I4c0c832da016b61266aaef9f4d40d40d504a4d0d
Reviewed-on: https://go-review.googlesource.com/33414
Reviewed-by: Than McIntosh <thanm@google.com>