runtime: move schedt type and sched var from C to Go
This doesn't change any actual code, it just starts using the Go
definition of the schedt type and the sched variable rather than the C
definitions.

The schedt type is tweaked slightly for gccgo.  We aren't going to
release goroutine stacks, so we don't need separate gfreeStack and
gfreeNostack lists.  We only have one size of defer function, so we
don't need a list of 5 different pools.

Change-Id: I372b9660cf64a0061e540a835000588648b36705
Reviewed-on: https://go-review.googlesource.com/33364
Reviewed-by: Than McIntosh <thanm@google.com>