runtime: fixes for -buildmode=c-archive
With -buildmode=c-archive, initsig is called before the memory
allocator has been initialized.  The code was doing a memory
allocation because of the call to funcPC(sigtramp).  When escape
analysis is fully implemented, that call should not allocate.  For
now, finesse the issue by calling a C function to get the C function
pointer value of sigtramp.

When returning from a call from C to a Go function, a deferred
function is run to go back to syscall mode.  When the call occurs on a
non-Go thread, that call sets g to nil, making it impossible to add
the _defer struct back to the pool.  Just drop it and let the garbage
collector clean it up.

Change-Id: I40c16ccb7463e103379a07093430323dcc9cd68c
Reviewed-on: https://go-review.googlesource.com/33675
Reviewed-by: Than McIntosh <thanm@google.com>