runtime: print C functions in traceback
Since gccgo can trace back through C code as easily as Go code, we
should print C functions in the traceback.

This worked before https://golang.org/cl/31230 for a dumb reason.  The
default value for runtime.traceback_cache was, and is, 2 << 2, meaning
to print all functions.  The old C code for runtime_parsedebugvars
would return immediately and do nothing if the environment variable
GODEBUG was not set (if GODEBUG was set it would later call
setTraceback.  The new Go code for runtime.parsedebugvars does not
return immediately if GODEBUG is not set, and always calls
setTraceback.  Either way, if GOTRACEBACK is not set, setTraceback
would set traceback_cache to 1 << 2, meaning to only print non-runtime
functions and having the effect of not printing plain C functions.

Keep the current handling of GODEBUG/GOTRACEBACK, which matches the gc
library, but add an extra check to print C functions by default.

Change-Id: I7a2219ee90c054b1e9b89bc5032eb1b0f0d375a3
Reviewed-on: https://go-review.googlesource.com/33717
Reviewed-by: Than McIntosh <thanm@google.com>