runtime: set isarchive in initsig
The library initialization code in go-libmain.c sets the C variable
runtime_isarchive but failed to set the Go variable runtime.isarchive.
We don't currently have a way to let C code access an unexported Go
variable, but fortunately the only time the Go function initsig is
called with an argument of true is exactly where we want to set
isarchive.  So let initsig do it.

Change-Id: Ic803b13b6899ca0856e0eccf2d4bfbe34860809c
Reviewed-on: https://go-review.googlesource.com/33753
Reviewed-by: Than McIntosh <thanm@google.com>