compiler: add slice initializers to the GC root list
As of https://golang.org/cl/32917 we can put slice initializers in the
.data section.  The program can still change the values in those
slices.  That means that if the slice elements can contain pointers,
we need to register the entire initializer as a GC root.

This would be straightforward except that we only have a Bexpression
for the slice initializer, not an Expression.  So introduce a
Backend_expression type that wraps a Bexpression as an Expression.

The test case for this is https://golang.org/cl/33790.

Change-Id: I183fa1e5cefe1a01a2a3fb6eff5167df938a34f7
Reviewed-on: https://go-review.googlesource.com/33792
Reviewed-by: Than McIntosh <thanm@google.com>