compiler: pass lvalue/rvalue context to back end for var exprs
Add a new flag on the Var_expression class that indicates
whether the var reference appears in an "lvalue" context
(for example, on the LHS of an assignment stmt) or an
"rvalue" context (for example, as an argument of a call).

Add a traversal pass that visits assignment stmt LHS subtrees
so as to mark things prior to backend gen. Select the right
context value in other places where Backend::var_expression is
called.

Change-Id: Id3c40b2898a645f81a6ae79cf83303eecc15473f
Reviewed-on: https://go-review.googlesource.com/33990
Reviewed-by: Ian Lance Taylor <iant@golang.org>