compiler: make Slice_construction_expression::do_flatten idempotent
Because of the way we handle call expressions with multiple results,
it's possible for expressions to be flattened more than once.  In the
case of Slice_construction_expression, allocating the slice storage
multiple times caused a compiler crash as one of the
Temporary_statement's wound up not getting a backend expression.

Test case is https://golang.org/cl/34020.

Change-Id: I2e1785092a5df717f02570d5a600e555994b202e
Reviewed-on: https://go-review.googlesource.com/34026
Reviewed-by: Than McIntosh <thanm@google.com>