compiler: add containing Bfunction to some backend interfaces.
Change the interfaces for backend methods that create statements to
always pass in the enclosing Bfunction for the statement.  Having the
function available simplifies things if a temporary variable has to be
created during the construction of a statement.

This also includes a change to the Mark_lvalue_varexprs helper
class to handle indirections on the left hand side of assignments
(e.g. "*x.y = ...").

Change-Id: I9e3ef0caa5089d46584678c83468ac3ef593617d
Reviewed-on: https://go-review.googlesource.com/34471
Reviewed-by: Ian Lance Taylor <iant@golang.org>