libgo: build with -Wa,-nH if possible on Solaris
By default the Solaris assembler records the required hardware
capability in the object file.  This means that the AES hashing code
breaks on systems that do not support AES, even though the code uses a
runtime check to only actually invoke the AES instructions on systems
that support it.  An earlier fix for the problem only fixed the shared
library, not the static libgo.a.  Fix the problem for real by using an
assembler option to not record the hardware capability.

For GCC PR 78978.

Patch by Rainer Orth.

Change-Id: I9343c4260001db9273f40b455cb40901c963cadb
Reviewed-on: https://go-review.googlesource.com/34910
Reviewed-by: Ian Lance Taylor <iant@golang.org>