compiler: mark generated struct/array types as incomparable
The recent change to generate type functions for more types with
identity comparisons caused us to generate some unnecessary functions,
and even caused a compiler crash on Solaris due to phase ordering.
Avoid this by marking all generated and uncompared struct and array
types as incomparable, so that we don't try to generate type functions
for them.

Change-Id: Ifbf5f3d2018a059bc5a0f6fb8baed24a926a2317
Reviewed-on: https://go-review.googlesource.com/35110
Reviewed-by: Than McIntosh <thanm@google.com>