compiler: implement type aliases
This is a start of implementing type aliases (`type T1 = T2`) in the
Go frontend.  This is incomplete, in that the reflection information
is not updated for an embedded type alias.  It is also not well
tested.  Finally, the change to the language has not been approved.
This should be regarded as preliminary work for experimental use.

Update golang/go#18130.

Change-Id: I5de44e4aa56088e0b1ee9d9b5ad2fbc5be40aa07
Reviewed-on: https://go-review.googlesource.com/35120
Reviewed-by: Ian Lance Taylor <iant@golang.org>