compiler: insert additional conversion for type desc ptr expr
Change the method Type::type_descriptor_pointer to apply an additional
type conversion to its result Bexpression, to avoid type clashes in
the back end. The backend expression for a given type descriptor var
is given a type of "_type", however the virtual calls that create the
variable use types derived from _type, hence the need to force a
conversion.

Change-Id: Iaa18348b34ec080b102abd1217b1ab3c512373ec
Reviewed-on: https://go-review.googlesource.com/35506
Reviewed-by: Ian Lance Taylor <iant@golang.org>