compiler: insure tree integrity in Call_expression::set_result
Depending on the back end, it can be problematic to reuse Bexpressions
(passing the same Bexpression to more than one Backend call to create
additional Bexpressions or Bstatements). The Call_expression::set_result
method was reusing its Bexpression input in more than one tree
context; the fix is to pass in an Expression instead and generate
multiple Bexpression references to it within the method.

Change-Id: Ic8c42aa2c698be9f1aa9260baa1944bcc456ac1a
Reviewed-on: https://go-review.googlesource.com/35505
Reviewed-by: Ian Lance Taylor <iant@golang.org>