runtime: fix context used by getTraceback
I typoed the argument passed to getcontext in getTraceback, and the
error was hidden by ucontext_arg.  This would have been caught by some
of the runtime package tests, but we don't run most of them because
they rely on `go build`, and the go tool is not available while
running the libgo testsuite.  We should fix the libgo testsuite,
somehow, so that they run.

Change-Id: I4f4be2044dbf30c9440be72317bc736364af8830
Reviewed-on: https://go-review.googlesource.com/35837
Reviewed-by: Than McIntosh <thanm@google.com>