runtime: fix setting of isarchive
When I updated to Go 1.8, the initsig function moved from
signal1_unix.go to signal_unix.go, and lost a gccgo-specific change in
the move.  Add it back.

This would have been caught by the misc/cgo/testcarchive tests in the
gc repository, but we don't run those in the gccgo repository.  We
should fix that, somehow.

Change-Id: I9c9f694bec60ab89b243dc3bdd971b0750648677
Reviewed-on: https://go-review.googlesource.com/35839
Reviewed-by: Than McIntosh <thanm@google.com>