compiler: let initializers hold addresses of fields
The runtime.dbgvars initializer looks like

var dbgvars = []dbgVar{
	{"allocfreetrace", &debug.allocfreetrace},
}

Because the field address was not recognized as valid for a static
initializer, the variable was initialized at runtime.  Normally that
would be fine, but for the runtime package it meant that dbgvars was
not initialized when it was read by parsedebugvars.  That meant that
the GODEBUG environment variable did nothing.

Fixing that revealed that the static initializer checks had to be more
careful about interface types, just like the existing is_constant
methods.

Change-Id: I7e39550431648941e501fdb26c6272d066f86ede
Reviewed-on: https://go-review.googlesource.com/35891
Reviewed-by: Than McIntosh <thanm@google.com>