libgo: update to final Go 1.8 release
Along with the update this fixes a problem that was always present but
only showed up with the new reflect test.  When a program used a
**unsafe.Pointer and stored the value in an interface type, the
generated type descriptor pointed to the GC data for *unsafe.Pointer.
It did that by name, but we were not generating a variable with the
right name.

Change-Id: I3f1060d9a8d9af207ec6e945425251edbecc410b
Reviewed-on: https://go-review.googlesource.com/37144
Reviewed-by: Than McIntosh <thanm@google.com>