libgo: use an import list for missing symbols
libgo depends on symbols provided by Go programs at runtime. On AIX,
this requires either to build libgo with -Wl,-berok linker option and
the programs with -Wl,-brtl, or to provide a list of imported symbols
when building libgo. The second options seems preferable, to avoid
requiring an additional option for every Go program.

There are also some symbols that are specific to GNU ld and do not
exist when linking with AIX ld (__data_start, __edata, __etext and
__bss_start).

Issue golang/go#19200

Change-Id: I2458460ebe7d096eeb0aa4cb15baed178ef117be
Reviewed-on: https://go-review.googlesource.com/37969
Reviewed-by: Ian Lance Taylor <iant@golang.org>