libgo: adapt sigtab.go to AIX
On AIX, _NSIG is not directly defined to its integer value in
gen-sysinfo.go.
The real value is _SIGMAX32+1 or _SIGMAX64+1, depending if we are
building a 32bit ligbo or a 64bit libgo, so we need to read one of
those constants to set nsig value in mksigtab.sh

This change also ensures that all signal numbers from 0 to nsig-1
are referenced in sigtable.

Change-Id: I479dfa6804647bc495b792d0be4a9698a297e96e
Reviewed-on: https://go-review.googlesource.com/38268
Reviewed-by: Ian Lance Taylor <iant@golang.org>